Moved cafe menu creation to app.py and added terminal color theming
